Update project package
Update project package name and structure to avoid future name collisions on github repository.
