+++ src/main/java/com/github/jiefenn8/graphloom/configmap/PredicateMap.java
@@ -1,0 +1,42 @@
+package com.github.jiefenn8.graphloom.configmap;
+
+/**
+ * Object Map
+ * <p>
+ * This interface defines the base methods that manages the mapping
+ * configuration of predicate between an entity and object specified
+ * by the {@link ObjectMap}.
+ */
+public interface PredicateMap {
+
+    /**
+     * Returns the predicate String used to describe the predicate (relationship)
+     * between the parent entity and object.
+     *
+     * @return the String of the predicate name
+     */
+    String getPredicate();
+
+    /**
+     * Sets the String to use for predicate URI generation and to describe
+     * the relationship between the parent entity and objects.
+     *
+     * @param predicate the predicate String to use for generation
+     */
+    void setPredicate(String predicate);
+
+    /**
+     * Returns the ObjectMap containing the configurations to retrieve object
+     * information to bind predicate generated by this {@code PredicateMap}
+     *
+     * @return {@code ObjectMap} containing the configs to retrieve object
+     */
+    ObjectMap getObjectMap();
+
+    /**
+     * @param objectMap the {@code ObjectMap} storing the object configs
+     *                  associated with this predicate map
+     */
+    void setObjectMap(ObjectMap objectMap);
+
+}
