--- src/main/java/r2graph/exceptions/InvalidMappingDocumentException.java
@@ -1,12 +1,0 @@
-package r2graph.exceptions;
-
-import r2graph.exceptions.base.FeijoaException;
-
-public class InvalidMappingDocumentException extends FeijoaException {
-    /**
-     * The MappingDocument is not valid.
-     */
-    public InvalidMappingDocumentException(String document) {
-        super(document);
-    }
-}
