--- src/main/java/r2graph/io/InputDatabase.java
@@ -1,8 +1,0 @@
-package r2graph.io;
-
-import java.util.List;
-import java.util.Map;
-
-public interface InputDatabase {
-    List<Map<String, String>> getRows(String table);
-}
