+++ src/test/java/com/github/jiefenn8/graphloom/r2rml/PredicateObjectMapObjectTest.java
@@ -1,0 +1,12 @@
+package com.github.jiefenn8.graphloom.r2rml;
+
+import com.github.jiefenn8.graphloom.configmap.AbstractObjectMapTest;
+import com.github.jiefenn8.graphloom.configmap.ObjectMap;
+
+public class PredicateObjectMapObjectTest extends AbstractObjectMapTest {
+
+    @Override
+    public ObjectMap createInstance() {
+        return new PredicateObjectMap();
+    }
+}
