--- src/test/java/r2graph/r2rml/PredicateObjectMapObjectTest.java
@@ -1,12 +1,0 @@
-package r2graph.r2rml;
-
-import r2graph.configmap.AbstractObjectMapTest;
-import r2graph.configmap.ObjectMap;
-
-public class PredicateObjectMapObjectTest extends AbstractObjectMapTest {
-
-    @Override
-    public ObjectMap createInstance() {
-        return new PredicateObjectMap();
-    }
-}
